Displaying currently logged in user and keeping the chat-bot state using context
